import_qpm_to_neo4j/requirements.txt
neo4j>=5.0.0
rdflib>=6.0.0
shapely>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0